Start handling partial conflict resolutions
These happen when 2 or 3 files conflict, and some of them are still
shown as conflicting, or the original file hasn't been crushed.

In this case, the system considers that the user is still undecided
until no conflict file remain (or the original file has been
overwritten).

This patch adds the few initial pieces of work to ensure partially
resolved conflict don't get to be undone when syncing with peers.